Move `float_pow` tests to their own file
